Changed product price to float

+++ database/migrations/2022_10_29_152334_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('price');
+            $table->decimal('price', 6, 2);
             $table->integer('quantity');
             $table->text('description');
             $table->foreignIdFor(Category::class);
